+++ sdk/core/System.ClientModel/src/ModelReaderWriter/ModelReaderWriter.cs
@@ -154,7 +154,7 @@ public static class ModelReaderWriter
         return model;
     }
 
-    private static object GetObjectInstance([DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.PublicConstructors | DynamicallyAccessedMemberTypes.NonPublicConstructors)] Type returnType)
+    internal static object GetObjectInstance([DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.PublicConstructors | DynamicallyAccessedMemberTypes.NonPublicConstructors)] Type returnType)
     {
         PersistableModelProxyAttribute? attribute = Attribute.GetCustomAttribute(returnType, typeof(PersistableModelProxyAttribute), false) as PersistableModelProxyAttribute;
         Type typeToActivate = attribute is null ? returnType : attribute.ProxyType;