+++ sdk/openai/Azure.AI.OpenAI/src/Custom/Assistants/AzureAssistantClient.cs
@@ -28,6 +28,5 @@ internal partial class AzureAssistantClient : AssistantClient
     }
 
     protected AzureAssistantClient()
-    {
-    }
+    { }
 }