+++ sdk/openai/Azure.AI.OpenAI/src/Custom/Audio/AzureAudioClient.cs
@@ -32,6 +32,5 @@ internal partial class AzureAudioClient : AudioClient
     }
 
     protected AzureAudioClient()
-    {
-    }
+    { }
 }