+++ sdk/openai/Azure.AI.OpenAI/src/Custom/Batch/AzureBatchClient.cs
@@ -32,6 +32,5 @@ internal partial class AzureBatchClient : BatchClient
     }
 
     protected AzureBatchClient()
-    {
-    }
+    { }
 }