+++ sdk/openai/Azure.AI.OpenAI/src/Custom/Chat/AzureChatClient.cs
@@ -35,8 +35,7 @@ internal partial class AzureChatClient : ChatClient
     }
 
     protected AzureChatClient()
-    {
-    }
+    { }
 
     /// <inheritdoc/>
     public override AsyncCollectionResult<StreamingChatCompletionUpdate> CompleteChatStreamingAsync(IEnumerable<ChatMessage> messages, ChatCompletionOptions options = null, CancellationToken cancellationToken = default)