+++ sdk/openai/Azure.AI.OpenAI/src/Custom/Embeddings/AzureEmbeddingClient.cs
@@ -32,6 +32,5 @@ internal partial class AzureEmbeddingClient : EmbeddingClient
     }
 
     protected AzureEmbeddingClient()
-    {
-    }
+    { }
 }