+++ sdk/openai/Azure.AI.OpenAI/src/Custom/FineTuning/AzureFineTuningClient.cs
@@ -29,6 +29,5 @@ internal partial class AzureFineTuningClient : FineTuningClient
     }
 
     protected AzureFineTuningClient()
-    {
-    }
+    { }
 }