+++ sdk/openai/Azure.AI.OpenAI/src/Custom/Images/AzureImageClient.cs
@@ -32,6 +32,5 @@ internal partial class AzureImageClient : ImageClient
     }
 
     protected AzureImageClient()
-    {
-    }
+    {}
 }