+++ sdk/openai/Azure.AI.OpenAI/src/Custom/VectorStores/AzureVectorStoreClient.cs
@@ -31,6 +31,5 @@ internal partial class AzureVectorStoreClient : VectorStoreClient
     }
 
     protected AzureVectorStoreClient()
-    {
-    }
+    { }
 }