+++ sdk/storage/Azure.Storage.DataMovement/api/Azure.Storage.DataMovement.net6.0.cs
@@ -195,10 +195,10 @@ namespace Azure.Storage.DataMovement
     {
         protected StorageResourceItemProperties() { }
         public StorageResourceItemProperties(long? resourceLength, Azure.ETag? eTag, System.DateTimeOffset? lastModifiedTime, System.Collections.Generic.Dictionary<string, object> properties) { }
-        public Azure.ETag? ETag { get { throw null; } }
-        public System.DateTimeOffset? LastModifiedTime { get { throw null; } }
-        public System.Collections.Generic.Dictionary<string, object> RawProperties { get { throw null; } }
-        public long? ResourceLength { get { throw null; } }
+        public Azure.ETag? ETag { get { throw null; } set { } }
+        public System.DateTimeOffset? LastModifiedTime { get { throw null; } set { } }
+        public System.Collections.Generic.Dictionary<string, object> RawProperties { get { throw null; } set { } }
+        public long? ResourceLength { get { throw null; } set { } }
     }
     public abstract partial class StorageResourceProvider
     {