+++ sdk/core/System.ClientModel/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 1.1.0-beta.7 (Unreleased)
+## 1.1.0-beta.8 (Unreleased)
 
 ### Features Added
 
@@ -10,7 +10,18 @@
 
 ### Other Changes
 
+- Removed implicit cast from `string` to `ApiKeyCredential`.
+
+## 1.1.0-beta.7 (2024-08-14)
+
+### Features Added
+
+- Added `JsonModelConverter` to allow integration with System.Text.Json.
+
+### Other Changes
+
 - Removed `ReturnWhen` enum in favor of using bool `waitUntilCompleted` parameter in third-party client LRO method signatures.
+- Added abstract `UpdateStatus` method to `OperationResult`.
 
 ## 1.1.0-beta.6 (2024-08-01)